Fix: config 폴더 삭제, config 파일 전부 디렉토리 최상단으로 이동
config 파일이 디렉토리 최상단에 없을 경우, 각 라이브러리가 프로젝트 설정을 찾지 못해 해당 설정에 따라 webpack에서 코드를 분석할 수 없는 문제가 생겨,
tsconfig.json, tailwind.config.js, posts.config.js 파일 모두 config 디렉토리 하단에서 디렉토리 최상단으로 이동.
