+++ package.json
@@ -19,7 +19,7 @@
     "react-redux": "^9.1.2",
     "react-router": "^6.24.1",
     "react-router-dom": "^6.24.1",
-    "react-scripts": "^5.0.1",
+    "react-scripts": "^3.0.1",
     "typescript": "^3.4.5",
     "web-vitals": "^2.1.4"
   },
