--- src/config/postcss.config.js
@@ -1,9 +1,0 @@
-const autoprefixer = require('autoprefixer');
-const tailwindConfig = require('./tailwind.config');
-
-module.exports = {
-  plugins: {
-    tailwindcss: { config: './tailwind.config.js' },
-    autoprefixer: {},
-  },
-};
